refactor(qbits): pack inside constructor
