+++ quanto/tensor/__init__.py
@@ -14,7 +14,6 @@
 
 from .core import *
 from .optimizers import *
-from .packed import *
 from .qactivation import *
 from .qbits import *
 from .qbytes import *
