+++ test/library/test_unpack.py
@@ -18,7 +18,7 @@ import pytest
 import torch
 
 from quanto.library import disable_extensions
-from quanto.tensor.packed import pack_weights
+from quanto.tensor.qbits.packed import pack_weights
 
 
 @pytest.mark.parametrize("bits", [2, 4], ids=["int2", "int4"])
